Move _page_utils, split out charts from components
