+++ vizro-core/examples/_chart-gallery/pages/magnitude.py
@@ -1,7 +1,7 @@
 from pages._factories import column_factory, treemap_factory
 import vizro.models as vm
 import vizro.plotly.express as px
-from utils._page_utils import DATA_DICT, PAGE_GRID, make_code_clipboard_from_py_file
+from pages._pages_utils import DATA_DICT, PAGE_GRID, make_code_clipboard_from_py_file
 
 # TODO: this is currently identical to ordered bar. It should be:
 #  - unordered
