--- vizro-core/examples/_chart-gallery/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Utils folder to contain helper functions and custom charts/components."""
